requirements.txt
# Core scientific computing
numpy>=1.20.0
scipy>=1.7.0

# FEniCS/DOLFINx for finite element simulation
fenics-dolfinx>=0.7.0
ufl>=2023.2.0
basix>=0.7.0

# Parallel computing
mpi4py>=3.1.0
petsc4py>=3.18.0

# Data handling and visualization
pandas>=1.3.0
matplotlib>=3.5.0

# Optional: For advanced mesh handling
meshio>=5.0.0

# Optional: For configuration management
pyyaml>=6.0